+++ lib/ftpd/ftp_server.rb
@@ -54,7 +54,6 @@ module Ftpd
 
     def session(socket)
       Session.new(:socket => socket,
-                  :interface => interface,
                   :driver => @driver,
                   :debug => @debug,
                   :debug_path => debug_path,